Move User from auth to users module
