+++ operationsgateway_api/src/routes/auth.py
@@ -5,10 +5,10 @@ from fastapi.responses import JSONResponse
 
 from operationsgateway_api.src.auth.authentication import Authentication
 from operationsgateway_api.src.auth.jwt_handler import JwtHandler
-from operationsgateway_api.src.auth.user import User
 from operationsgateway_api.src.error_handling import endpoint_error_handling
 from operationsgateway_api.src.exceptions import ForbiddenError
 from operationsgateway_api.src.models import AccessTokenModel, LoginDetailsModel
+from operationsgateway_api.src.users.user import User
 
 
 log = logging.getLogger()
